feat: DrumPadContainer finished
This component holds a list of sound files and a list of letters.
The letters represent the keys and the sound files are the ones to be
played by the drum machine. It generates the drumpad by making use of
the Array.map() function in the letters array. It creates a DrumPad
component for each letter, passing said letter as id and the link at the
same index as the letter on the links array